+++ source/renderer/app/components/staking/widgets/TooltipPool.js
@@ -40,10 +40,13 @@ import {
   ARROW_HEIGHT,
   ARROW_OFFSET,
   TOOLTIP_DELTA,
+  LIST_VIEW_TOOLTIP_DELTA_TOP,
+  LIST_VIEW_ROW_HEIGHT,
   TOOLTIP_MAX_HEIGHT,
   TOOLTIP_WIDTH,
   IS_RANKING_DATA_AVAILABLE,
   IS_SATURATION_DATA_AVAILABLE,
+  THUMBNAIL_WIDTH,
 } from '../../../config/stakingConfig';
 
 const messages = defineMessages({
@@ -166,6 +169,7 @@ const messages = defineMessages({
 type Props = {
   stakePool: StakePool,
   isVisible: boolean,
+  fromStakePool?: boolean,
   currentTheme: string,
   onClick: Function,
   onOpenExternalLink: Function,
@@ -176,6 +180,7 @@ type Props = {
   color: string,
   containerClassName: string,
   numberOfRankedStakePools: number,
+  isListView: boolean,
 };
 
 type State = {
@@ -269,9 +274,9 @@ export default class TooltipPool extends Component<Props, State> {
     } =
       tooltipPosition === 'top' || tooltipPosition === 'bottom'
         ? this.getTopBottomPosition(left)
-        : this.getLeftRightPosition(top, isTopHalf);
+        : this.getLeftRightPosition(top, isTopHalf, left);
 
-    const componentStyle = this.getComponenStyle(
+    const componentStyle = this.getComponentStyle(
       tooltipPosition,
       componentTop,
       componentBottom,
@@ -296,6 +301,7 @@ export default class TooltipPool extends Component<Props, State> {
   };
 
   getTopBottomPosition = (left: number) => {
+    const { fromStakePool, isListView } = this.props;
     const paddingOffset = rangeMap(
       left,
       THUMBNAIL_OFFSET_WIDTH,
@@ -304,14 +310,25 @@ export default class TooltipPool extends Component<Props, State> {
       THUMBNAIL_OFFSET_WIDTH / 2
     );
 
-    const componentLeft =
-      -((TOOLTIP_WIDTH * left) / this.containerWidth) +
-      THUMBNAIL_OFFSET_WIDTH +
-      paddingOffset;
-    const componentTop = THUMBNAIL_HEIGHT + ARROW_HEIGHT / 2;
-    const componentBottom = THUMBNAIL_HEIGHT + ARROW_HEIGHT / 2;
+    const componentLeft = !fromStakePool
+      ? -((TOOLTIP_WIDTH * left) / this.containerWidth) +
+        THUMBNAIL_OFFSET_WIDTH +
+        paddingOffset
+      : -((TOOLTIP_WIDTH * left) / this.containerWidth) +
+        THUMBNAIL_OFFSET_WIDTH +
+        (left - THUMBNAIL_OFFSET_WIDTH - THUMBNAIL_WIDTH);
+    let componentTop = !fromStakePool
+      ? THUMBNAIL_HEIGHT + ARROW_HEIGHT / 2
+      : THUMBNAIL_HEIGHT + ARROW_WIDTH / 2;
+    if (isListView) componentTop -= LIST_VIEW_ROW_HEIGHT;
+    let componentBottom = !fromStakePool
+      ? THUMBNAIL_HEIGHT + ARROW_HEIGHT / 2
+      : THUMBNAIL_HEIGHT / 2;
+    if (isListView) componentBottom += ARROW_HEIGHT;
 
-    const arrowLeft = -componentLeft + THUMBNAIL_OFFSET_WIDTH - ARROW_OFFSET;
+    const arrowLeft = !fromStakePool
+      ? -componentLeft + THUMBNAIL_OFFSET_WIDTH - ARROW_OFFSET
+      : THUMBNAIL_HEIGHT - ARROW_WIDTH - TOOLTIP_DELTA;
     const arrowTop = -(ARROW_WIDTH / 2);
     const arrowBottom = -(ARROW_WIDTH / 2);
 
@@ -325,21 +342,26 @@ export default class TooltipPool extends Component<Props, State> {
     };
   };
 
-  getLeftRightPosition = (top: number, isTopHalf: boolean) => {
+  getLeftRightPosition = (top: number, isTopHalf: boolean, left: number) => {
+    const { fromStakePool, isListView } = this.props;
     const bottom = this.containerHeight - (top + THUMBNAIL_HEIGHT);
-
-    const componentLeft = THUMBNAIL_HEIGHT;
+    const componentLeft = fromStakePool
+      ? -((TOOLTIP_WIDTH * left) / this.containerWidth) +
+        THUMBNAIL_OFFSET_WIDTH +
+        (left - THUMBNAIL_OFFSET_WIDTH + ARROW_HEIGHT)
+      : THUMBNAIL_HEIGHT;
     let componentTop = 'auto';
     let componentBottom = 'auto';
     let arrowTop = 'auto';
     let arrowBottom = 'auto';
 
-    if (isTopHalf) {
+    if (isTopHalf && !isListView) {
       componentTop = -((TOOLTIP_MAX_HEIGHT * top) / this.containerHeight);
       arrowTop = -componentTop + ARROW_WIDTH / 2;
     } else {
       componentBottom = -((TOOLTIP_MAX_HEIGHT * bottom) / this.containerHeight);
       arrowBottom = -componentBottom + ARROW_WIDTH / 2;
+      if (fromStakePool) arrowBottom -= TOOLTIP_DELTA;
     }
 
     const arrowLeft = -(ARROW_WIDTH / 2);
@@ -355,17 +377,16 @@ export default class TooltipPool extends Component<Props, State> {
   };
 
   getTooltipPosition = (top: number, isLeftHalf: boolean) => {
-    const ignoreTopBottom = false;
-    if (!ignoreTopBottom) {
-      if (top <= TOOLTIP_DELTA) {
-        return 'bottom';
-      }
-      if (
-        TOOLTIP_DELTA >=
-        this.containerHeight - (top + (THUMBNAIL_HEIGHT - TOOLTIP_DELTA))
-      ) {
-        return 'top';
-      }
+    const { isListView } = this.props;
+    const topDelta = isListView ? LIST_VIEW_TOOLTIP_DELTA_TOP : TOOLTIP_DELTA;
+    if (top <= topDelta) {
+      return 'bottom';
+    }
+    if (
+      TOOLTIP_DELTA >=
+      this.containerHeight - (top + (THUMBNAIL_HEIGHT - TOOLTIP_DELTA))
+    ) {
+      return 'top';
     }
     if (!isLeftHalf) {
       return 'left';
@@ -373,7 +394,7 @@ export default class TooltipPool extends Component<Props, State> {
     return 'right';
   };
 
-  getComponenStyle = (
+  getComponentStyle = (
     tooltipPosition: string,
     top: number | 'auto',
     bottom: number | 'auto',
@@ -480,11 +501,11 @@ export default class TooltipPool extends Component<Props, State> {
               <span className={saturationBarClassnames}>
                 <span
                   style={{
-                    width: `${parseFloat(saturation.toFixed(2))}%`,
+                    width: `${parseFloat(saturation).toFixed(2)}%`,
                   }}
                 />
               </span>
-              {`${parseFloat(saturation.toFixed(2))}%`}
+              {`${parseFloat(saturation).toFixed(2)}%`}
             </span>
           </div>
         ),
@@ -534,8 +555,8 @@ export default class TooltipPool extends Component<Props, State> {
         value: (
           <div className={styles.defaultColor}>
             <span className={styles.defaultColorContent}>{`${parseFloat(
-              relativeStake.toFixed(2)
-            )}%`}</span>
+              relativeStake
+            ).toFixed(2)}%`}</span>
           </div>
         ),
       },
@@ -551,7 +572,7 @@ export default class TooltipPool extends Component<Props, State> {
                 }),
               }}
             >
-              {`${parseFloat(profitMargin.toFixed(2))}%`}
+              {`${parseFloat(profitMargin).toFixed(2)}%`}
             </span>
           </div>
         ),